Refactored the pulling the UserLocation coordinates, and it is now in the AppDelegate so it is the first thing that is calculated when the app starts.
