+++ Cheers/Controller/AppDelegate.swift
@@ -7,17 +7,22 @@
 //
 
 import UIKit
+import CoreLocation
 
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate {
 
     var window: UIWindow?
+    var userLocation: CLLocationCoordinate2D? //Holds UserLocation Coordinates
+    var locationManager: CLLocationManager!
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
+        getUserLocation(); //Makes a call to determine user location on launch
+        
         return true
     }
+    
 
     func applicationWillResignActive(_ application: UIApplication) {
         // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
@@ -40,6 +45,37 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationWillTerminate(_ application: UIApplication) {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
     }
+    
+    
+    
+//    //MARK: - Getting User Location
+//    /***************************************************************/
+    
+    //Helper function
+    func getUserLocation() {
+        initLocationManager();
+    }
+    
+    //inits all the required LocationManager settings
+    func initLocationManager() {
+        locationManager = CLLocationManager()
+        locationManager.delegate = self
+        CLLocationManager.locationServicesEnabled()
+        locationManager.desiredAccuracy = kCLLocationAccuracyKilometer
+        locationManager.requestWhenInUseAuthorization()
+        locationManager.startUpdatingLocation()
+    }
+    
+    //Pulls latest location and stops updating.
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        let locationArray = locations as NSArray
+        let locationObj = locationArray.lastObject as? CLLocation
+        let coord = locationObj?.coordinate
+        userLocation = coord
+        locationManager.stopUpdatingLocation()
+        
+    }
+    
 
 
 }
