+++ Cheers/Controller/MapViewController.swift
@@ -8,26 +8,23 @@
 
 import UIKit
 import MapKit
-import CoreLocation
 
-class MapViewController: UIViewController, CLLocationManagerDelegate {
+class MapViewController: UIViewController {
     @IBOutlet weak var Map: MKMapView!
     
     var barList: [Place] = []
     var myLocation: CLLocationCoordinate2D?
-    let locationManager = CLLocationManager()
+    
+    //used to pull info from AppDelegate
+    let appDelegate = UIApplication.shared.delegate as! AppDelegate
+    
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        //Current Location required set-up
-        locationManager.delegate = self
-        locationManager.requestWhenInUseAuthorization()
-        locationManager.desiredAccuracy = kCLLocationAccuracyKilometer
-        locationManager.startUpdatingLocation()
 
-        // set up my location
-        myLocation = CLLocationCoordinate2DMake(32.774063, -117.185914)
+        //Gets the current location from the caculation in the AppDelegate
+        myLocation = appDelegate.userLocation
+        Map.showsUserLocation = true
 
         // load in bars
         populateMap()
@@ -35,27 +32,6 @@ class MapViewController: UIViewController, CLLocationManagerDelegate {
         
     }
     
-    
-    //Reads the location manager for latest coordinate and displays it on map.
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        
-        let location = locations[0]
-        
-        
-        myLocation = CLLocationCoordinate2DMake(location.coordinate.latitude , location.coordinate.longitude)
-
-        
-        let center = location.coordinate
-        let span = MKCoordinateSpan(latitudeDelta: 0.7, longitudeDelta: 0.7)
-        let reigon = MKCoordinateRegion(center: center, span: span)
-        
-        Map.setRegion(reigon, animated: true)
-        Map.showsUserLocation = true
-        
-        locationManager.stopUpdatingLocation()
-        
-    }
-  
     
     func populateMap() {
         // this call will be replaced by API call once core features are functioning
